Check the Operation for the field is a Mutation
