+++ layers/Engine/packages/component-model/src/StandaloneCheckpoints/EnabledMutationsCheckpoint.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoP\ComponentModel\StandaloneCheckpoints;
+
+use PoP\ComponentModel\FeedbackItemProviders\CheckpointErrorFeedbackItemProvider;
+use PoP\ComponentModel\Module;
+use PoP\ComponentModel\ModuleConfiguration;
+use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
+use PoP\GraphQLParser\Spec\Parser\Ast\MutationOperation;
+use PoP\GraphQLParser\Spec\Parser\Ast\OperationInterface;
+use PoP\Root\App;
+use PoP\Root\Feedback\FeedbackItemResolution;
+
+class EnabledMutationsCheckpoint extends AbstractStandaloneCheckpoint
+{
+    private ?CheckpointErrorFeedbackItemProvider $checkpointErrorFeedbackItemProvider = null;
+
+    final public function setCheckpointErrorFeedbackItemProvider(CheckpointErrorFeedbackItemProvider $checkpointErrorFeedbackItemProvider): void
+    {
+        $this->checkpointErrorFeedbackItemProvider = $checkpointErrorFeedbackItemProvider;
+    }
+    final protected function getCheckpointErrorFeedbackItemProvider(): CheckpointErrorFeedbackItemProvider
+    {
+        /** @var CheckpointErrorFeedbackItemProvider */
+        return $this->checkpointErrorFeedbackItemProvider ??= $this->instanceManager->getInstance(CheckpointErrorFeedbackItemProvider::class);
+    }
+
+    public function __construct(
+        protected FieldInterface $field,
+    ) {        
+    }
+
+    public function validateCheckpoint(): ?FeedbackItemResolution
+    {
+        /** @var ModuleConfiguration */
+        $moduleConfiguration = App::getModule(Module::class)->getConfiguration();
+        if (!$moduleConfiguration->enableMutations()) {
+            return new FeedbackItemResolution(
+                CheckpointErrorFeedbackItemProvider::class,
+                CheckpointErrorFeedbackItemProvider::E1
+            );
+        }
+
+        /**
+         * Get the Operation for the field, and check
+         * that it is a Mutation
+         *
+         * @var SplObjectStorage<AstInterface,AstInterface>
+         */
+        $documentASTNodeAncestors = App::getState('document-ast-node-ancestors');
+        $astNode = $this->field;
+        $astNodeTopMostAncestor = null;
+        while ($astNode !== null) {
+            $astNodeTopMostAncestor = $astNode;
+            $astNode = $documentASTNodeAncestors[$astNode] ?? null;
+        }
+        if ($astNodeTopMostAncestor instanceof OperationInterface) {
+            $operation = $astNodeTopMostAncestor;
+            if (!($operation instanceof MutationOperation)) {
+                return new FeedbackItemResolution(
+                    CheckpointErrorFeedbackItemProvider::class,
+                    CheckpointErrorFeedbackItemProvider::E1
+                );
+            }
+        }
+
+        return parent::validateCheckpoint();
+    }
+}
